+++ resources/release-notes.md
@@ -6,4 +6,5 @@
 
 - (Bartington, HelioSwarm, IMAP) Add results preview for science data
 - (Bartington, HelioSwarm, IMAP) Increase app area size by 30 %
+- (Bartington, HelioSwarm, IMAP) Exported results are named after the selected mission (e.g., `imapAnalysis` instead of `analysis`)
 - (IMAP) Show any available metadata for sensor setup